convert_bigquery_data/requirements.txt
google-cloud-storage==1.35.0
google-auth==1.24.0
PyYAML==5.3.1
ndjson==0.3.1